sftBack/sftMadness/src/requirements.txt
psycopg2-binary==2.9.10
requests==2.31.0
boto3==1.29.0
botocore>=1.31.44
PyJWT[crypto]==2.8.0
cryptography==36.0.0
botocore>=1.31.44
python-dotenv
pymongo
beautifulsoup4==4.12.3